Develop 1

+++ src/main/java/com/chan/stock_portfolio_backtest_api/repository/CalcIndexPriceRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface CalcIndexPriceRepository extends JpaRepository<CalcIndexPrice, Long> {
+public interface CalcIndexPriceRepository extends JpaRepository<CalcIndexPrice, Integer> {
     List<CalcIndexPrice> findByIndexInfoAndBaseDateBetween(IndexInfo indexInfo, LocalDate startDate, LocalDate endDate);
 }
